+++ frontend/src/components/organisms/InstanceTable.tsx
@@ -1,4 +1,5 @@
 import { Button, ButtonGroup, Code, HTMLTable, Intent, NonIdealState, Spinner } from "@blueprintjs/core";
+import { IconNames } from "@blueprintjs/icons";
 import { push } from "connected-react-router";
 import { range, sortBy, sortedUniq, zip } from "lodash";
 import * as numeral from "numeral";
@@ -7,7 +8,7 @@ import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import styled from "styled-components";
 import { loadInstanceList } from "../../redux/actions";
-import { IAppState, IInstanceListResponse } from "../../redux/types";
+import { IAppState, IInstanceListResponse, IInstanceSort, SortField } from "../../redux/types";
 import { InstanceType } from "../atoms";
 import { ErrorState } from "../molecules";
 
@@ -21,19 +22,38 @@ const PaginationContainer = styled.div`
   flex: 1;
   align-items: center;
 `;
+const InstanceColumn = styled.th`
+  width: 15%;
+`;
+const ServerColumn = styled.th`
+  width: 20%;
+`;
+const VersionColumn = styled.th`
+  width: 20%;
+`;
+const UserCountColumn = styled.th`
+  width: 15%;
+`;
+const StatusCountColumn = styled.th`
+  width: 15%;
+`;
+const InsularityColumn = styled.th`
+  width: 15%;
+`;
 
 interface IInstanceTableProps {
   loadError: boolean;
   instancesResponse?: IInstanceListResponse;
+  instanceListSort: IInstanceSort;
   isLoading: boolean;
-  fetchInstances: (page?: number) => void;
+  loadInstanceList: (page?: number, sort?: IInstanceSort) => void;
   navigate: (path: string) => void;
 }
 class InstanceTable extends React.PureComponent<IInstanceTableProps> {
   public componentDidMount() {
     const { isLoading, instancesResponse, loadError } = this.props;
     if (!isLoading && !instancesResponse && !loadError) {
-      this.props.fetchInstances();
+      this.props.loadInstanceList();
     }
   }
 
@@ -53,12 +73,44 @@ class InstanceTable extends React.PureComponent<IInstanceTableProps> {
         <StyledTable striped={true} bordered={true} interactive={true}>
           <thead>
             <tr>
-              <th>Instance</th>
-              <th>Server type</th>
-              <th>Version</th>
-              <th>Users</th>
-              <th>Statuses</th>
-              <th>Insularity</th>
+              <InstanceColumn>
+                Instance
+                <Button
+                  minimal={true}
+                  icon={this.getSortIcon("domain")}
+                  onClick={this.sortByFactory("domain")}
+                  intent={this.getSortIntent("domain")}
+                />
+              </InstanceColumn>
+              <ServerColumn>Server type</ServerColumn>
+              <VersionColumn>Version</VersionColumn>
+              <UserCountColumn>
+                Users
+                <Button
+                  minimal={true}
+                  icon={this.getSortIcon("userCount")}
+                  onClick={this.sortByFactory("userCount")}
+                  intent={this.getSortIntent("userCount")}
+                />
+              </UserCountColumn>
+              <StatusCountColumn>
+                Statuses
+                <Button
+                  minimal={true}
+                  icon={this.getSortIcon("statusCount")}
+                  onClick={this.sortByFactory("statusCount")}
+                  intent={this.getSortIntent("statusCount")}
+                />
+              </StatusCountColumn>
+              <InsularityColumn>
+                Insularity
+                <Button
+                  minimal={true}
+                  icon={this.getSortIcon("insularity")}
+                  onClick={this.sortByFactory("insularity")}
+                  intent={this.getSortIntent("insularity")}
+                />
+              </InsularityColumn>
             </tr>
           </thead>
           <tbody>
@@ -67,8 +119,8 @@ class InstanceTable extends React.PureComponent<IInstanceTableProps> {
                 <td>{i.name}</td>
                 <td>{i.type && <InstanceType type={i.type} />}</td>
                 <td>{i.version && <Code>{i.version}</Code>}</td>
-                <td>{i.userCount}</td>
-                <td>{i.statusCount}</td>
+                <td>{i.userCount && numeral.default(i.userCount).format("0,0")}</td>
+                <td>{i.statusCount && numeral.default(i.statusCount).format("0,0")}</td>
                 <td>{i.insularity && numeral.default(i.insularity).format("0.0%")}</td>
               </tr>
             ))}
@@ -84,7 +136,7 @@ class InstanceTable extends React.PureComponent<IInstanceTableProps> {
           <ButtonGroup>
             {zip(pagesToDisplay, pagesToDisplay.slice(1)).map(([page, nextPage], idx) => {
               if (page === undefined) {
-                return;
+                return null;
               }
               const isCurrentPage = currentPage === page;
               const isEndOfSection = nextPage !== undefined && page + 1 !== nextPage && page !== totalPages;
@@ -113,12 +165,42 @@ class InstanceTable extends React.PureComponent<IInstanceTableProps> {
     );
   }
 
+  private sortByFactory = (field: SortField) => () => {
+    const { instancesResponse, instanceListSort } = this.props;
+
+    const page = (instancesResponse && instancesResponse.pageNumber) || 1;
+    const nextSortDirection =
+      instanceListSort.field === field && instanceListSort.direction === "desc" ? "asc" : "desc";
+
+    this.props.loadInstanceList(page, { field, direction: nextSortDirection });
+  };
+
   private loadPageFactory = (page: number) => () => {
-    this.props.fetchInstances(page);
+    this.props.loadInstanceList(page);
   };
 
   private goToInstanceFactory = (domain: string) => () => {
     this.props.navigate(`/instance/${domain}`);
+  };
+
+  private getSortIcon = (field: SortField) => {
+    const { instanceListSort } = this.props;
+    if (instanceListSort.field !== field) {
+      return IconNames.SORT;
+    } else if (instanceListSort.direction === "asc") {
+      return IconNames.SORT_ASC;
+    } else {
+      return IconNames.SORT_DESC;
+    }
+  };
+
+  private getSortIntent = (field: SortField) => {
+    const { instanceListSort } = this.props;
+    if (instanceListSort.field === field) {
+      return Intent.PRIMARY;
+    } else {
+      return Intent.NONE;
+    }
   };
 
   private getPagesToDisplay = (totalPages: number, currentPage: number) => {
@@ -138,13 +220,14 @@ class InstanceTable extends React.PureComponent<IInstanceTableProps> {
 
 const mapStateToProps = (state: IAppState) => {
   return {
+    instanceListSort: state.data.instanceListSort,
     instancesResponse: state.data.instancesResponse,
     isLoading: state.data.isLoadingInstanceList,
     loadError: state.data.instanceListLoadError
   };
 };
 const mapDispatchToProps = (dispatch: Dispatch) => ({
-  fetchInstances: (page?: number) => dispatch(loadInstanceList(page) as any),
+  loadInstanceList: (page?: number, sort?: IInstanceSort) => dispatch(loadInstanceList(page, sort) as any),
   navigate: (path: string) => dispatch(push(path))
 });